[PR Feedback] Merge nats module into event module
